Added randomized waves and changed scaling
with arrays for each wave 1 - 10 and an 11th wave that is very hard for all waves above the first ten. scales the amount of enemy groups each wave and picks a random set of groups each time. also changed scaling to a function that graphs up at a good pace to make the game stay challenging.

made bounty scale with difficulty, now that difficulty scales much more than before