Fixed pause menu bug, added wavemax control
on the pause menu players can set the max waves. it is synced between them. when one player joins, it updates the max so they are synced up. when the set max is reached, you win. max possible waves is 99.

fixed a bug that caused players to sometimes get launched forever by bullets and just drift along