v1.0 early access beta
Added tutorial screen at the beginning of the game.

Fixed bug that caused set resources not to be synced properly.

Fixed bug that caused wave to start early if enemy count dropped to zero before respawning (zombie enemy).

Lowered health for a few enemies.

Current known bugs: none (but there probably are some)

Assuming I don't find any obvious bugs, this is going to be the last update to this project for a while so I can move onto other projects.